fix: overhaul printer
Fix patterns so the end of a pattern exactly matches the end index of a word
(previously the end was bigger than the actuall end by 1). In this regard change
ranges in print functions.

Change the sorting comparator so it compares the ends of patterns if their starts
are equal.

Interface out print functions so they read from a io.Reader and write to io.Writer.
Add wrapper functions to pass io.Reader and set io.Writer to os.Stdout.

Extract hardcoded input name placeholder to constant.